requirements.txt
django==4.2.10
djangorestframework==3.14.0
markdown==3.5
django-filter==23.5
pillow==10.2.0
coreapi==2.3.3
pyyaml==6.0.1
uritemplate==4.1.1 